# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Socrates API",
    "version": "2026.05",
    "description": "Socrates (topodrive.top) is an AI tutor. This document covers the public API surface — health, bootstrap, OAuth 2.0 authorization-code flow, scoped agent API keys, async import jobs, and a bulk-create endpoint. Authenticated browser routes (`/api/sessions`, `/api/chat`, `/api/files`) follow the same conventions and reuse the schemas declared here. Long-form reference: https://topodrive.top/llms-full.txt. Agent instructions: https://topodrive.top/agents.md.",
    "contact": {
      "name": "Socrates",
      "email": "help@addtech.site",
      "url": "https://topodrive.top/contact"
    },
    "license": {
      "name": "Proprietary"
    },
    "x-deprecation-policy": "Versioned under `/api/v2/*`. Deprecations signal via the `Deprecation` and `Sunset` response headers (RFC 8594) at least 90 days before removal. See https://topodrive.top/announcements for the announcement feed."
  },
  "servers": [
    {
      "url": "https://app.topodrive.top",
      "description": "Production"
    },
    {
      "url": "https://app.topodrive.top/api/v2",
      "description": "Production (mobile-shell contract prefix; nginx rewrites /api/v2/* → /api/* on the same origin)"
    }
  ],
...
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "operationId": "authLogin",
        "summary": "Email/password login",
        "description": "Authenticate a browser user via email and password. Issues a `sid` session cookie on success.",
        "security": [{ "csrfDoubleSubmit": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/LoginRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful. Set-Cookie: sid=<64-hex>; HttpOnly; SameSite=Lax.",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SessionInfo" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    },
...
    "/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "operationId": "authLogout",
        "summary": "Clear the session",
        "description": "Server-side revocation of the `sid` cookie. Idempotent — calling on a non-existent session returns 200 with `{ ok: true }`.",
        "security": [{ "csrfDoubleSubmit": [] }],
        "responses": {
          "200": {
            "description": "Session cleared.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["ok"],
                  "properties": { "ok": { "type": "boolean", "enum": [true] } }
                }
              }
            }
          },
          "403": { "$ref": "#/components/responses/Forbidden" }
        }
      }
    },
...
    "/chat/stream": {
      "post": {
        "tags": ["Auth"],
        "operationId": "streamChatCompletion",
        "summary": "Streaming chat completion",
        "description": "Stream a Socratic tutor reply as `text/event-stream`. The tutor asks before it answers. Reuse the same `Idempotency-Key` to make network retries safe — replays return the same stream.",
        "security": [{ "sessionCookie": [], "oauthBearer": [] }],
        "parameters": [
          { "$ref": "#/components/parameters/IdempotencyKey" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/ChatStreamRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Server-Sent Events stream of the tutor's reply.",
            "content": {
              "text/event-stream": {
                "schema": { "$ref": "#/components/schemas/ChatStreamEvent" }
              }
            },
            "headers": {
              "RateLimit-Limit": { "$ref": "#/components/headers/RateLimitLimit" },
              "RateLimit-Remaining": { "$ref": "#/components/headers/RateLimitRemaining" },
              "RateLimit-Reset": { "$ref": "#/components/headers/RateLimitReset" }
            }
          },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "429": { "$ref": "#/components/responses/TooManyRequests" }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sid",
        "description": "Browser session cookie issued by /api/auth/login. 64-hex opaque token, HttpOnly, SameSite=Lax."
      },
      "mobileBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "ma.<pairId>.<64-hex>",
        "description": "Mobile access token issued by /api/mobile/login. 15-minute TTL."
      },
      "mobileRefresh": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Mobile refresh token: `Authorization: Bearer mr.<pairId>.<64-hex>`."
      },
      "csrfDoubleSubmit": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "CSRF double-submit token. Pair with a `csrf-token` cookie issued by /api/auth/csrf-token."
      },
      "oauthBearer": {
        "type": "oauth2",
        "description": "OAuth 2.0 access token issued by /api/oauth/token, or a scoped agent API key (ak_*). Carries one or more of the scopes below.",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://app.topodrive.top/api/oauth/authorize",
            "tokenUrl": "https://app.topodrive.top/api/oauth/token",
            "refreshUrl": "https://app.topodrive.top/api/oauth/token",
            "scopes": {
              "chat:read": "Read your chat sessions and messages.",
              "chat:write": "Send messages on your behalf.",
              "memory:read": "Read your cross-session memory notes.",
              "memory:write": "Add or update your cross-session memory notes.",
              "sessions:read": "List your tutoring sessions and their metadata.",
              "sessions:write": "Create or modify tutoring sessions.",
              "files:read": "Download files you uploaded or that Socrates produced for you.",
...
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": false,
        "schema": {
          "type": "string",
          "maxLength": 255,
          "description": "Client-generated unique key (UUID recommended). Replays within the 24-hour window return the original response."
        },
        "description": "Make this write idempotent: same key + same request = same response, without duplicating side effects. Use on every mutating call when the network is unreliable."
      },
...
    },
    "headers": {
      "RateLimitLimit": {
        "description": "The maximum number of requests permitted in the current window.",
        "schema": { "type": "integer" }
      },
      "RateLimitRemaining": {
        "description": "The number of requests remaining in the current window.",
        "schema": { "type": "integer" }
      },
      "RateLimitReset": {
        "description": "Seconds until the current rate-limit window resets.",
        "schema": { "type": "integer" }
      },
      "RetryAfter": {
        "description": "RFC 6585 — seconds until the client should retry. Present on 429 responses.",
        "schema": { "type": "integer" }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Request was malformed. See `code` for the machine-readable reason.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid credentials.",
        "headers": {
          "WWW-Authenticate": {
            "description": "RFC 6750 challenge. Includes `resource_metadata` so the agent can resolve the full auth chain from a single 401.",
            "schema": { "type": "string" }
          }
        },
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "Forbidden": {
        "description": "Credentials are valid but insufficient. Includes `error=\"insufficient_scope\"` if a scope check failed.",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
...
      "TooManyRequests": {
        "description": "Rate limit hit. Respect `Retry-After` and back off.",
        "headers": {
          "Retry-After": { "$ref": "#/components/headers/RetryAfter" },
          "RateLimit-Limit": { "$ref": "#/components/headers/RateLimitLimit" },
          "RateLimit-Remaining": { "$ref": "#/components/headers/RateLimitRemaining" },
          "RateLimit-Reset": { "$ref": "#/components/headers/RateLimitReset" }
        },
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": ["code", "message"],
        "description": "Typed error envelope. `code` is a stable machine-readable identifier from a closed vocabulary; `message` is a one-sentence explanation safe to show to a user; `detail` carries optional structured context; `requestId` correlates to the `X-Request-Id` response header.",
        "properties": {
          "code": {
            "type": "string",
            "description": "Machine-readable error code (e.g. `BAD_REQUEST`, `UNAUTHORIZED`, `INSUFFICIENT_SCOPE`, `TOO_MANY_REQUESTS`).",
            "example": "BAD_REQUEST"
          },
          "message": {
            "type": "string",
            "description": "Human-readable one-sentence explanation.",
            "example": "source and payload required"
          },
          "detail": {
            "description": "Optional structured context for the error — field-level validation errors, retry timing, etc.",
            "oneOf": [
              { "type": "object", "additionalProperties": true },
              { "type": "array", "items": { "type": "object", "additionalProperties": true } }
            ]
          },
          "requestId": {
            "type": "string",
            "description": "Server-generated UUID. Matches the `X-Request-Id` response header for cross-log correlation.",
            "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479"
          },
          "retryAfterSeconds": {
            "type": "integer",
            "description": "Present on 429 responses. Seconds until the next request will be accepted.",
            "example": 60
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "format": "password", "minLength": 8 }
        }
      },
...
      "SessionInfo": {
        "type": "object",
        "required": ["ok", "user"],
        "properties": {
          "ok": { "type": "boolean" },
          "user": {
            "type": "object",
            "required": ["id", "email"],
            "properties": {
              "id": { "type": "string" },
              "email": { "type": "string", "format": "email" }
            }
          }
        }
      },
...
      "ChatStreamRequest": {
        "type": "object",
        "required": ["sessionId", "content"],
        "properties": {
          "sessionId": { "type": "string", "format": "uuid" },
          "content": { "type": "string", "minLength": 1, "maxLength": 32000 },
          "mode": { "type": "string", "enum": ["socratic", "tutor"], "default": "socratic" },
          "response_speed": { "type": "string", "enum": ["standard", "fast"], "default": "standard", "description": "Requests priority service when fast; unsupported providers automatically fall back to standard." }
        }
      },
      "ChatStreamEvent": {
        "type": "object",
        "description": "Single SSE frame. Wire format: `event: <event>\\ndata: <json>\\n\\n`.",
        "required": ["event", "data"],
        "properties": {
          "event": {
            "type": "string",
            "enum": ["start", "token", "tool", "preference_fallback", "complete", "error"]
          },
          "data": { "type": "object", "additionalProperties": true }
        }
      }
    }
  }
}